move var and std to fluid/layers/nn.py test=release/2.0
